+++ data/tasks.txt
@@ -1,8 +1,6 @@
-D | 1 | submit hw | 29/10/2024 1800
 E | 0 | project meeting | 10/10/2024 1700 | 10/10/2024 1800
-D | 1 | complete week 4 | 10/12/2024 1500
 E | 1 | football | 10/10/2024 1800 | 10/10/2024 1900
-T | 1 | go gym
 D | 0 | eat dinner | 14/8/2024 1900
 E | 0 | swimming | 12/3/2024 0800 | 12/3/2024 1200
 E | 0 | fow | 26/7/2024 0800 | 29/7/2024 2200
+T | 0 | gym
