+++ src/main/java/myapp/command/MarkCommand.java
@@ -1,7 +1,14 @@
-public class UnMarkCommand extends Command {
+package myapp.command;
+
+import myapp.core.BingBongUI;
+import myapp.core.Storage;
+import myapp.task.TaskList;
+import myapp.task.Task;
+
+public class MarkCommand extends Command {
 
     int index;
-    public UnMarkCommand(int i) {
+    public MarkCommand(int i) {
         super();
         this.index = i;
     }
@@ -13,9 +20,9 @@ public class UnMarkCommand extends Command {
             throw new IndexOutOfBoundsException();
         }
         Task task = tasks.get(index);
-        task.markAsNotDone();
+        task.markAsDone();
         saveTasks(tasks, ui, storage);
-        ui.showResponse("OK, I've marked this task as not done yet:\n" + task);
+        ui.showResponse("Nice! I've marked this task as done:\n" + task);
     }
 
     @Override
