+++ src/main/java/myapp/core/BingBongException.java
@@ -1,3 +1,5 @@
+package myapp.core;
+
 public class BingBongException extends Exception {
     public BingBongException(String message) {
         super(message);
