+++ src/main/java/myapp/task/Task.java
@@ -1,3 +1,5 @@
+package myapp.task;
+
 abstract public class Task {
     protected String description;
     protected boolean isDone;
