+++ src/main/java/myapp/task/Todo.java
@@ -1,3 +1,5 @@
+package myapp.task;
+
 public class Todo extends Task{
 
     public Todo(String description) {
